+++ insights-ui/src/components/ticker/debug/DebugTickerDividends.tsx
@@ -6,55 +6,53 @@ import Accordion from '@dodao/web-core/utils/accordion/Accordion';
 import getBaseUrl from '@dodao/web-core/utils/api/getBaseURL';
 import { useState } from 'react';
 import ConfirmationModal from '@dodao/web-core/components/app/Modal/ConfirmationModal';
-import TickerFinancialsButton from './TickerFinancialsButton';
 import { safeParseJsonString } from '@/util/safe-parse-json-string';
 
-export interface DebugTickerFinancialsProps {
+export interface DebugTickerDividendsProps {
   report: FullNestedTickerReport;
   onPostUpdate: () => Promise<void>;
 }
 
-export default function DebugTickerFinancials({ report, onPostUpdate }: DebugTickerFinancialsProps) {
+export default function DebugTickerDividends({ report, onPostUpdate }: DebugTickerDividendsProps) {
   const ticker = report.tickerKey;
 
-  const tickerFinancials = safeParseJsonString(report.tickerInfo).financials ?? '';
+  const tickerDividends = safeParseJsonString(report.tickerInfo).dividends ?? '';
 
   const [selectedCriterionAccordian, setSelectedCriterionAccordian] = useState<string | null>(null);
   const [showConfirmModal, setShowConfirmModal] = useState(false);
 
   const {
-    postData: regenerateTickerFinancials,
-    loading: TickerFinancialsLoading,
-    error: TickerFinancialsError,
+    postData: regenerateTickerDividends,
+    loading: TickerDividendsLoading,
+    error: TickerDividendsError,
   } = usePostData<string, {}>({
-    errorMessage: 'Failed to repopulate ticker financials.',
+    errorMessage: 'Failed to repopulate ticker dividends.',
   });
 
-  const handleRegenerateTickerFinancials = async () => {
-    await regenerateTickerFinancials(`${getBaseUrl()}/api/tickers/${ticker}/ticker-financials`);
+  const handleRegenerateTickerDividends = async () => {
+    await regenerateTickerDividends(`${getBaseUrl()}/api/tickers/${ticker}/ticker-dividends`);
     await onPostUpdate();
   };
 
   return (
     <div className="mt-8">
-      {TickerFinancialsError && <div className="text-red-500">{TickerFinancialsError}</div>}
+      {TickerDividendsError && <div className="text-red-500">{TickerDividendsError}</div>}
       <PrivateWrapper>
         <div className="flex justify-end">
-          <Button loading={TickerFinancialsLoading} primary variant="contained" onClick={() => setShowConfirmModal(true)} disabled={TickerFinancialsLoading}>
-            Repopulate Ticker Financials
+          <Button loading={TickerDividendsLoading} primary variant="contained" onClick={() => setShowConfirmModal(true)} disabled={TickerDividendsLoading}>
+            Repopulate Ticker Dividends
           </Button>
         </div>
       </PrivateWrapper>
-      <h1 className="mb-8 font-bold text-xl">Ticker Financials</h1>
+      <h1 className="mb-8 font-bold text-xl">Ticker Dividends</h1>
       <Accordion
-        label={'Ticker Financials'}
-        isOpen={selectedCriterionAccordian === `ticker_financials`}
-        onClick={() => setSelectedCriterionAccordian(selectedCriterionAccordian === `ticker_financials` ? null : `ticker_financials`)}
+        label={'Ticker Dividends'}
+        isOpen={selectedCriterionAccordian === `ticker_dividends`}
+        onClick={() => setSelectedCriterionAccordian(selectedCriterionAccordian === `ticker_dividends` ? null : `ticker_dividends`)}
       >
         <div className="mt-4">
-          <TickerFinancialsButton tickerKey={ticker} tickerFinancialsContent={JSON.stringify(tickerFinancials, null, 2) || undefined} onUpdate={onPostUpdate} />
           <pre className="whitespace-pre-wrap break-words overflow-x-auto max-h-[400px] overflow-y-auto text-xs">
-            {tickerFinancials ? JSON.stringify(tickerFinancials, null, 2) : 'Not populated yet'}
+            {tickerDividends ? JSON.stringify(tickerDividends, null, 2) : 'Not populated yet'}
           </pre>
         </div>
       </Accordion>
@@ -63,13 +61,13 @@ export default function DebugTickerFinancials({ report, onPostUpdate }: DebugTic
           open={showConfirmModal}
           onClose={() => setShowConfirmModal(false)}
           onConfirm={async () => {
-            await handleRegenerateTickerFinancials();
+            await handleRegenerateTickerDividends();
             setShowConfirmModal(false);
           }}
-          title="Repopulate Ticker Financials"
-          confirmationText="Are you sure you want to repopulate the ticker financials?"
+          title="Repopulate Ticker Dividends"
+          confirmationText="Are you sure you want to repopulate the ticker dividends?"
           askForTextInput={true}
-          confirming={TickerFinancialsLoading}
+          confirming={TickerDividendsLoading}
         />
       )}
     </div>
